// src/shared/model/CollegeId.ts
export enum CollegeId {
  AI = "CLG00001",
  DT = "CLG00002",
  HAPPINESS = "CLG00003",
  SV = "CLG00004",
  DESIGN = "CLG00005",
  GLOBAL = "CLG00006",
  LEADERSHIP = "CLG00007",
  MANAGEMENT = "CLG00008",
  SEMICONDUCTOR = "CLG00019",
  SKMANAGEMENT = "CLG00017",
  SKACADEMY = "CLG00018",
  LIFESTYLE = "CLG0001a",
  ENERGYSOLUTION = "CLG0001c",
  SKIS_COMMON = "CLG0001w",
  BM_DESIGNER = "CLG00020",
}